refactor(score-calc): 編成スロットを DeckSlots に抽出 (D&D 含む Runes 化)
